KeyLogger class
