+++ src/extension.ts
@@ -6,25 +6,27 @@ import { SubscriptionService } from './services/subscriptionService';
 import { CommandCounter } from './main/counter/commandCounter';
 import * as platform from './main/platform';
 import { KeybindingStorage } from './main/keybindings/keybindings';
+import { KeyLogger } from './main/keylogging/KeyLogger';
 
 // This method is called when your extension is activated
 // Your extension is activated the very first time the command is executed
 export function activate(context: vscode.ExtensionContext) {
+	let keyLogger = new KeyLogger();
 	uIOhook.on('keydown', (e) => {
-		commandCounter.handleKeyDown(e.keycode);
+		keyLogger.handleKeyDown(e.keycode);
 	});
 	uIOhook.on('keyup', (e) => {
-		commandCounter.handleKeyUp(e.keycode);
+		keyLogger.handleKeyUp(e.keycode);
 	});
 	uIOhook.on('mousedown', (_) => {
-		commandCounter.handleMousePress();
+		keyLogger.handleMousePress();
 	});
 	uIOhook.start();
 
 	console.log('Reading keybindings...');
 	const keybindingStorage = new KeybindingStorage(platform.get());
 	console.log(keybindingStorage.getKeybindingMap());
-	const commandCounter = new CommandCounter(0, keybindingStorage);
+	const commandCounter = new CommandCounter(0, keybindingStorage, keyLogger);
 	const subscriptionService = new SubscriptionService(commandCounter);
 	subscriptionService.listenForPossibleShortcutActions();
