+++ src/main/keylogging/KeyLogger.test.ts
@@ -1,0 +1,6 @@
+import * as assert from 'assert';
+import { KeyLogger } from "./KeyLogger";
+
+describe("Key Logger Test", () => {
+
+});
